+++ src/main/java/xyz/theprogramsrc/supercoreapi/spigot/guis/BrowserGUI.java
@@ -1,14 +1,15 @@
 package xyz.theprogramsrc.supercoreapi.spigot.guis;
 
-import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import xyz.theprogramsrc.supercoreapi.global.translations.Base;
 import xyz.theprogramsrc.supercoreapi.spigot.dialog.Dialog;
 import xyz.theprogramsrc.supercoreapi.spigot.guis.action.ClickAction;
+import xyz.theprogramsrc.supercoreapi.spigot.guis.events.GUIEvent;
+import xyz.theprogramsrc.supercoreapi.spigot.guis.events.GUIOpenEvent;
 import xyz.theprogramsrc.supercoreapi.spigot.guis.objects.GUIRows;
 import xyz.theprogramsrc.supercoreapi.spigot.items.SimpleItem;
 
-import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -16,6 +17,7 @@ public abstract class BrowserGUI<OBJ> extends GUI {
 
     private String searchTerm = null;
     public int maxItemsPerPage = 36;
+    public boolean backEnabled = false;
     private int page = 0;
 
     /**
@@ -32,39 +34,33 @@ public abstract class BrowserGUI<OBJ> extends GUI {
     }
 
     @Override
-    public GUIButton[] getButtons() {
-        OBJ[] objs = this.getObjects();
-        List<OBJ> objectsFound = new ArrayList<>();
-        for (OBJ obj : objs) {
-            if(obj != null){
-                if(this.searchTerm == null){
-                    objectsFound.add(obj);
-                }else{
-                    GUIButton button = this.getButton(obj);
-                    String name = ChatColor.stripColor(new SimpleItem(button.getItemStack()).getDisplayName()).toLowerCase();
-                    if(name.contains(ChatColor.stripColor(this.searchTerm.toLowerCase()))){
-                        objectsFound.add(obj);
-                    }
+    public void onEvent(GUIEvent event) {
+        if(event instanceof GUIOpenEvent){
+            this.clearButtons();
+            OBJ[] objs = this.getObjects();
+            List<OBJ> objectsFound = Arrays.stream(this.getObjects()).filter(obj->{
+                if(this.searchTerm != null){
+                    String itemName = this.getSuperUtils().removeColor(new SimpleItem(this.getButton(obj).getItemStack()).getDisplayName()).toLowerCase();
+                    String search = this.getSuperUtils().removeColor(this.searchTerm).toLowerCase();
+                    return itemName.contains(search);
                 }
-            }
-        }
-        int index0 = this.page * this.maxItemsPerPage;
-        int index1 = Math.min(index0 + this.maxItemsPerPage, objectsFound.size());
-        int maxPages = (int)Math.round(Math.ceil((double)objectsFound.size() / (double)maxItemsPerPage));
-        List<GUIButton> buttons1 = new ArrayList<>();
-        int slot = 0;
-        for (GUIButton b : objectsFound.subList(index0, index1).stream().map(this::getButton).collect(Collectors.toList())) {
-            b.setSlot(slot);
-            buttons1.add(b);
-            slot++;
-        }
-        List<GUIButton> buttons = new ArrayList<>(buttons1);
-        buttons1.clear(); // Save memory
-        objectsFound.clear(); // Save memory
 
-        buttons.add(new GUIButton(49, this.searchTerm == null ? this.getPreloadedItems().getSearchItem() : this.getPreloadedItems().getEndSearchItem()).setAction(a -> {
+                return true;
+            }).collect(Collectors.toList());
+            int index0 = this.page * this.maxItemsPerPage;
+            int index1 = Math.min(index0 + this.maxItemsPerPage, objectsFound.size());
+            int maxPages = (int)Math.round(Math.ceil((double)objectsFound.size() / (double)maxItemsPerPage));
+            List<GUIButton> buttons =  objectsFound.subList(index0, index1).stream().map(this::getButton).collect(Collectors.toList());
+            int i = 0;
+            for(int o = 0; o < buttons.size(); ++o){
+                buttons.set(o, buttons.get(o).setSlot(i));
+                i++;
+            }
+
+            this.addButtons(buttons);
+            this.removeButtons(43, 44,45,46,47,48,49,50,51,52,53);
             if(this.searchTerm == null){
-                new Dialog(this.getPlayer()){
+                this.addButton(new GUIButton(49, this.getPreloadedItems().getSearchItem(), a-> new Dialog(this.getPlayer()){
 
                     @Override
                     public String getTitle() {
@@ -88,32 +84,33 @@ public abstract class BrowserGUI<OBJ> extends GUI {
                     }
                 }.setRecall(player ->{
                     this.page = 0;
-                    this.open();
-                });
+                    this.getSpigotTasks().runTask(this::open);
+                })));
             }else{
-                this.searchTerm = null;
-                this.page = 0;
-                this.open();
+                this.addButton(new GUIButton(49, this.getPreloadedItems().getEndSearchItem(), a->{
+                    this.searchTerm = null;
+                    this.page = 0;
+                    this.open();
+                }));
             }
-        }));
-        buttons.add(new GUIButton(45, this.getPreloadedItems().getBackItem()).setAction(this::onBack));
 
-        if(this.page != 0){
-            buttons.add(new GUIButton(52, this.getPreloadedItems().getPreviousItem()).setAction(a -> {
-                this.page--;
-                this.open();
-            }));
-        }
-        if(this.page+1 < maxPages){
-            buttons.add(new GUIButton(53, this.getPreloadedItems().getNextItem()).setAction(a -> {
-                this.page++;
-                this.open();
-            }));
-        }
+            if(this.page != 0){
+                this.addButton(new GUIButton(52, this.getPreloadedItems().getPreviousItem()).setAction(a -> {
+                    this.page -= 1;
+                    this.open();
+                }));
+            }
+            if(this.page+1 < maxPages){
+                this.addButton(new GUIButton(53, this.getPreloadedItems().getNextItem()).setAction(a -> {
+                    this.page += 1;
+                    this.open();
+                }));
+            }
 
-        GUIButton[] guiButtons = new GUIButton[buttons.size()];
-        guiButtons = buttons.toArray(guiButtons);
-        return guiButtons;
+            if(this.backEnabled){
+                this.addButton(new GUIButton(45, this.getPreloadedItems().getBackItem()).setAction(this::onBack));
+            }
+        }
     }
 
     /**
@@ -133,5 +130,9 @@ public abstract class BrowserGUI<OBJ> extends GUI {
      * Executed when the back button is clicked
      * @param clickAction The ClickAction
      */
-    public abstract void onBack(ClickAction clickAction);
+    public void onBack(ClickAction clickAction){
+
+    }
+
+
 }