+++ src/main/java/xyz/theprogramsrc/supercoreapi/spigot/guis/events/GUIUpdateEvent.java
@@ -1,0 +1,11 @@
+package xyz.theprogramsrc.supercoreapi.spigot.guis.events;
+
+import xyz.theprogramsrc.supercoreapi.spigot.guis.GUI;
+
+public class GUIUpdateEvent extends GUIEvent{
+
+    public GUIUpdateEvent(GUI gui) {
+        super(gui);
+    }
+
+}